# Conflicts:
#	pom.xml
#	src/main/java/io/coti/cotinode/data/BaseTransactionData.java
#	src/main/java/io/coti/cotinode/http/websocket/UpdatedBalanceMessage.java
#	src/main/java/io/coti/cotinode/service/BalanceService.java

+++ src/main/java/io/coti/cotinode/http/GetBalancesResponse.java
@@ -3,13 +3,15 @@ package io.coti.cotinode.http;
 import io.coti.cotinode.data.Hash;
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.AbstractMap;
 import java.util.List;
 
 @Data
 public class GetBalancesResponse extends Response {
 
-    private List<AbstractMap.SimpleEntry<Hash, Double>> amounts;
+    private List<AbstractMap.SimpleEntry<Hash, BigDecimal>> amounts;
 
     public GetBalancesResponse() {
         super("Balance details");
